refactor: auth 관련 API 경로를 통일하여 구조 개선 (#65)
